refacto(ai_copilot): rename AiAgent to AiCopilot
