+++ ui/src/components/ai/AiCopilot.vue
